+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## WIP
+
+## Changed
+- clearer error messages for byte buffer overflow
+
 ## 0.2.15
 
 ## Fixed